# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão e Monitoramento de Casos de Cólera",
        "description": "Documentação da API para Sistema de Gestão e Monitoramento de Casos de Cólera em Angola",
        "contact": {
            "name": "Suporte Técnico",
            "email": "admin@saudeangola.gov"
        },
        "license": {
            "name": "Licença MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Versão 1"
        }
    ],
    "paths": {
...
        "/gabinetes": {
            "get": {
                "tags": [
                    "Gabinetes"
                ],
                "summary": "Listar todos os gabinetes provinciais",
                "description": "Retorna uma lista de todos os gabinetes provinciais cadastrados",
                "operationId": "listarGabinetesProvinciais",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gabinetes provinciais listados com sucesso."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Gabinete Provincial de Luanda"
                                                    },
                                                    "endereco": {
                                                        "type": "string",
                                                        "example": "Av. 4 de Fevereiro, 42"
                                                    },
                                                    "telefone": {
                                                        "type": "string",
                                                        "example": "+244 923456789"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "gabinete.luanda@saude.gov.ao"
                                                    },
                                                    "diretor": {
                                                        "type": "string",
                                                        "example": "Dr. João Silva"
                                                    },
                                                    "latitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": -8.838333
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 13.234444
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao listar gabinetes provinciais: [mensagem de erro]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gabinetes"
                ],
                "summary": "Criar um novo gabinete provincial",
                "description": "Cria um novo registro de gabinete provincial",
                "operationId": "criarGabineteProvincial",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "endereco",
                                    "telefone",
                                    "email",
                                    "diretor",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Gabinete Provincial de Benguela"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Rua Principal, 123"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "+244 923456789"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "gabinete.benguela@saude.gov.ao"
                                    },
                                    "diretor": {
                                        "type": "string",
                                        "example": "Dra. Maria Santos"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -12.5789
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 13.4071
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gabinete provincial criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gabinete provincial criado com sucesso."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GabineteProvincial"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Não autorizado a criar gabinetes provinciais"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de validação"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "GabineteProvincial": {
                "title": "Gabinete Provincial",
                "description": "Modelo de Gabinete Provincial",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Gabinete Provincial de Saúde de Luanda"
                    },
                    "provincia": {
                        "type": "string",
                        "example": "Luanda"
                    },
                    "endereco": {
                        "type": "string",
                        "example": "Av. Revolução de Outubro, 112"
                    },
                    "telefone": {
                        "type": "string",
                        "example": "+244 923456789"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "gabinete.luanda@saude.gov.ao"
                    },
                    "diretor": {
                        "type": "string",
                        "example": "Dr. António Santos"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -8.838333
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 13.234444
                    },
                    "ativo": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}